+++ src/utils.rs
@@ -2,3 +2,4 @@ pub mod timer;
 pub mod globals;
 pub mod machine;
 pub mod counter;
+pub mod tinkerer;